Preparing version 1.0.0
Problem: We need to have this library working across different versions
of Ruby and Rails. We will do a major release with the changes required
to achieve that goal.

Solution: This first commit sanitizes the dependencies (remove
development dependencies and declared previously hidden dependency upon
active_support). It also removed the mokeypatching on the String class.