Fix result consistency with Rails 6.1+
Problem: Since https://github.com/rails/rails/pull/32313,
`ActiveModel::Errors` does not hold anymore a hash of errors: it
introduced a new object `ActiveModel::Error` which
`ActiveModel::Errors` maintains a list of. Later,
`ActiveModel::Errors` was enriched with the `#errors` method that
return the list of structured errors.

As a result, calling `invalid!` with `ActiveModel::Error` will return
the list of structured error and no more the good old hashes of
errors. This is the case when the `request_object` is an invalid
`ActiveModel`.

Solution: To workaround this issue, we forbid the calls to `#errors`
on `ActiveModel::Errors` objects.